// dashboard/src/utils/hasura-api/generated/schemas/requestTransformation.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { RequestTransformationBody } from './requestTransformationBody';
import type { RequestTransformationContentType } from './requestTransformationContentType';
import type { RequestTransformationMethod } from './requestTransformationMethod';
import type { RequestTransformationQueryParams } from './requestTransformationQueryParams';
import type { RequestTransformationTemplateEngine } from './requestTransformationTemplateEngine';
import type { RequestTransformationVersion } from './requestTransformationVersion';
import type { TransformHeaders } from './transformHeaders';

/**
 * Request transformation
 */
export interface RequestTransformation {
  /** Sets the RequestTransformation schema version. Version 1 uses a String for the body field and Version 2 takes a BodyTransform. */
  version?: RequestTransformationVersion;
  /** Change the request method to this value. */
  method?: RequestTransformationMethod;
  /** Change the request URL to this value. */
  url?: string;
  /** Replace the Content-Type with this value. */
  content_type?: RequestTransformationContentType;
  /** Request Header Transformation. */
  request_headers?: TransformHeaders;
  /** Replace the query params on the URL with this value. You can specify a dictionary of key/value pairs which is converted into a query string or directly give a query string. */
  query_params?: RequestTransformationQueryParams;
  /** Template language to be used for this transformation. */
  template_engine?: RequestTransformationTemplateEngine;
  /** A template script for transforming the request body. */
  body?: RequestTransformationBody;
}